+++ CHANGELOG.md
@@ -36,8 +36,7 @@ To Document:
 
 TO DO:
 * IN PROGRESS: Factor out RunResult
-  * Convert beforeAll/afterAll from TestCaseResult[] to RunResult[]
-  * Convert TestResult.suite() to TestSuiteResult.create() and TestResult.testCase() to TestCaseResult.create()
+  * Convert TestResult.testCase() to TestCaseResult.create()
   * Revise TestSuiteResult.create() to match signature of other result factories
   * Revise TestRenderer to more specific types where possible (including RunResult instead of TestResult)
 * Check that stack trace highlighting works with beforeEach etc.